move fragment masking plugin to own file
